business_feasibility/requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.3.2
transformers==4.44.2
torch==2.4.1
numpy==1.26.4